Generative AI/Dynamic Pricing App/requirements.txt
dash
pandas 
scikit-learn 
python-dotenv 
google-generativeai
langgraph